📝 docs(deployments): update account and region configuration The account.hcl file in the root_spin-aws deployment was updated to change the account_name variable from "learn-aft" to "spin-aws". The aws_arn variable was also updated to reflect the new account. Additionally, a comment was added to the file with a link to documentation on how to add accounts/roles.
A new deployment was added in the deployments directory called "tf_testaccount". This deployment contains an account.hcl file with variables specific to the new account, as well as a config.yml file and a region.hcl file for the eu-west-2 region. A new VPC was also added in the eu-west-2 region with the name "sandbox-vpc-0". The VPC has a CIDR block of "10.1.0.0/24" and spans across three availability zones: eu-west-2a, eu-west-2b